+++ android/src/main/java/com/github/blueboytm/flutter_v2ray/v2ray/V2rayController.java
@@ -25,12 +25,19 @@ public class V2rayController {
         Utilities.copyAssets(context);
         AppConfigs.APPLICATION_ICON = app_icon;
         AppConfigs.APPLICATION_NAME = app_name;
-        context.registerReceiver(new BroadcastReceiver() {
+
+        BroadcastReceiver receiver = new BroadcastReceiver() {
             @Override
             public void onReceive(Context arg0, Intent arg1) {
                 AppConfigs.V2RAY_STATE = (AppConfigs.V2RAY_STATES) arg1.getExtras().getSerializable("STATE");
             }
-        }, new IntentFilter("V2RAY_CONNECTION_INFO"));
+        };
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+             context.registerReceiver(receiver, new IntentFilter("V2RAY_CONNECTION_INFO"), Context.RECEIVER_EXPORTED);
+        }else {
+             context.registerReceiver(receiver, new IntentFilter("V2RAY_CONNECTION_INFO"));
+        }
+        // context.registerReceiver(receiver, new IntentFilter("V2RAY_CONNECTION_INFO"), null, null);
     }
 
     public static void changeConnectionMode(final AppConfigs.V2RAY_CONNECTION_MODES connection_mode) {
@@ -111,8 +118,8 @@ public class V2rayController {
         return delay[0];
     }
 
-    public static long getV2rayServerDelay(final String config) {
-        return V2rayCoreManager.getInstance().getV2rayServerDelay(config);
+    public static long getV2rayServerDelay(final String config, final String url) {
+        return V2rayCoreManager.getInstance().getV2rayServerDelay(config, url);
     }
 
     public static AppConfigs.V2RAY_CONNECTION_MODES getConnectionMode() {