+++ android/src/main/java/com/github/blueboytm/flutter_v2ray/v2ray/core/V2rayCoreManager.java
@@ -76,10 +76,10 @@ public final class V2rayCoreManager {
                 Intent connection_info_intent = new Intent("V2RAY_CONNECTION_INFO");
                 connection_info_intent.putExtra("STATE", V2rayCoreManager.getInstance().V2RAY_STATE);
                 connection_info_intent.putExtra("DURATION", SERVICE_DURATION);
-                connection_info_intent.putExtra("UPLOAD_SPEED", Utilities.parseTraffic(uploadSpeed, false, true));
-                connection_info_intent.putExtra("DOWNLOAD_SPEED", Utilities.parseTraffic(downloadSpeed, false, true));
-                connection_info_intent.putExtra("UPLOAD_TRAFFIC", Utilities.parseTraffic(totalUpload, false, false));
-                connection_info_intent.putExtra("DOWNLOAD_TRAFFIC", Utilities.parseTraffic(totalDownload, false, false));
+                connection_info_intent.putExtra("UPLOAD_SPEED", uploadSpeed);
+                connection_info_intent.putExtra("DOWNLOAD_SPEED", downloadSpeed);
+                connection_info_intent.putExtra("UPLOAD_TRAFFIC", totalUpload);
+                connection_info_intent.putExtra("DOWNLOAD_TRAFFIC", totalDownload);
                 context.sendBroadcast(connection_info_intent);
             }
 
@@ -95,7 +95,7 @@ public final class V2rayCoreManager {
     public void setUpListener(Service targetService) {
         try {
             v2rayServicesListener = (V2rayServicesListener) targetService;
-            Libv2ray.initV2Env(getUserAssetsPath(targetService.getApplicationContext()));
+            Libv2ray.initV2Env(getUserAssetsPath(targetService.getApplicationContext()), "");
             isLibV2rayCoreInitialized = true;
             SERVICE_DURATION = "00:00:00";
             seconds = 0;
@@ -173,7 +173,7 @@ public final class V2rayCoreManager {
             stopCore();
         }
         try {
-            Libv2ray.testConfig(v2rayConfig.V2RAY_FULL_JSON_CONFIG);
+            // Libv2ray.testConfig(v2rayConfig.V2RAY_FULL_JSON_CONFIG);
         } catch (Exception e) {
             sendDisconnectedBroadCast();
             Log.e(V2rayCoreManager.class.getSimpleName(), "startCore failed => v2ray json config not valid.");
@@ -219,10 +219,10 @@ public final class V2rayCoreManager {
             Intent connection_info_intent = new Intent("V2RAY_CONNECTION_INFO");
             connection_info_intent.putExtra("STATE", V2rayCoreManager.getInstance().V2RAY_STATE);
             connection_info_intent.putExtra("DURATION", SERVICE_DURATION);
-            connection_info_intent.putExtra("UPLOAD_SPEED", Utilities.parseTraffic(0, false, true));
-            connection_info_intent.putExtra("DOWNLOAD_SPEED", Utilities.parseTraffic(0, false, true));
-            connection_info_intent.putExtra("UPLOAD_TRAFFIC", Utilities.parseTraffic(0, false, false));
-            connection_info_intent.putExtra("DOWNLOAD_TRAFFIC", Utilities.parseTraffic(0, false, false));
+            connection_info_intent.putExtra("UPLOAD_SPEED", uploadSpeed);
+            connection_info_intent.putExtra("DOWNLOAD_SPEED", uploadSpeed);
+            connection_info_intent.putExtra("UPLOAD_TRAFFIC", uploadSpeed);
+            connection_info_intent.putExtra("DOWNLOAD_TRAFFIC", uploadSpeed);
             try {
                 v2rayServicesListener.getService().getApplicationContext().sendBroadcast(connection_info_intent);
             } catch (Exception e) {
@@ -259,13 +259,13 @@ public final class V2rayCoreManager {
 
     public Long getConnectedV2rayServerDelay() {
         try {
-            return v2RayPoint.measureDelay();
+            return v2RayPoint.measureDelay(AppConfigs.DELAY_URL);
         } catch (Exception e) {
             return -1L;
         }
     }
 
-    public Long getV2rayServerDelay(final String config) {
+    public Long getV2rayServerDelay(final String config, final String url) {
         try {
             try {
                 JSONObject config_json = new JSONObject(config);
@@ -273,10 +273,10 @@ public final class V2rayCoreManager {
                 new_routing_json.remove("rules");
                 config_json.remove("routing");
                 config_json.put("routing", new_routing_json);
-                return Libv2ray.measureOutboundDelay(config_json.toString());
+                return Libv2ray.measureOutboundDelay(config_json.toString(), url);
             } catch (Exception json_error) {
                 Log.e("getV2rayServerDelay", json_error.toString());
-                return Libv2ray.measureOutboundDelay(config);
+                return Libv2ray.measureOutboundDelay(config, url);
             }
         } catch (Exception e) {
             Log.e("getV2rayServerDelayCore", e.toString());